Add getUpdates method
We were already providing the getUpdates method from the Bot class, but we were offering it in a different way than we do for all the other API methods (e.g. returning dispatchable objects instead of the real API result). We are adding the new alternative for flexibility (because it's not recommended to use this method but instead consume updates through the library built-in mechanisms). And it's the first step in the plan to remove the other getUpdates method we offer from the Bot class because it shouldn't be there and it's leaking implementation details.