Separate Dispatcher and Updater
Updater is the class that carries out the polling to get updates from Telegram server and put them into a queue of updates. Dispatcher is a class that consumes updates from this queue and calls to the required update handlers. Dispatcher was embeded within Updater and we are separating them so they are independent components.